requirements.txt

opencv-python==4.8.0.76
face-recognition==1.3.0
pillow==9.5.0
requests==2.31.0
